Change how graph displays its data based on if you choose one or two measurements

+++ Orkulausnir/Models/MeasurementItem.cs
@@ -4,8 +4,8 @@ namespace Orkulausnir.Models
     {
         public int Harmonic { get; set; }
 
-        public decimal Average { get; set; }
+        public double Average { get; set; }
 
-        public decimal Max { get; set; }
+        public double Max { get; set; }
     }
 }
